refactor(settings): extract document codec
