+++ src/main/settings/repository.test.ts
@@ -3,7 +3,8 @@ import { isAbsolute, join, normalize, sep } from 'node:path'
 import { tmpdir } from 'node:os'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { SettingsRepository, sanitizeSettings } from './repository'
+import { sanitizeSettings } from './document-codec'
+import { SettingsRepository } from './repository'
 import type { StoredProvider } from './types'
 
 // Capture the warn calls the repository makes through createLogger. vi.hoisted runs before the
@@ -1052,17 +1053,19 @@ describe('settings repository: v2 official providers & activeModel migration', (
     expect((await repository.getSettings()).notebookRuntimes).toBeUndefined()
   })
 
-  it('rejects a malformed runtime selection (no interpreter path)', async () => {
+  it('rejects malformed runtime selections before applying language constraints', async () => {
     const repository = new SettingsRepository(await createStorageRoot())
 
-    await expect(
-      repository.setRuntimeSelection('python', {
-        source: 'external',
-        interpreterPath: '',
-        appOwnedOverlay: false,
-        packageInstallAuthorized: false
-      })
-    ).rejects.toThrow(/invalid/i)
+    for (const language of ['python', 'r'] as const) {
+      await expect(
+        repository.setRuntimeSelection(language, {
+          source: 'external',
+          interpreterPath: '',
+          appOwnedOverlay: false,
+          packageInstallAuthorized: false
+        })
+      ).rejects.toThrow(/invalid/i)
+    }
   })
 
   it('persists and clears a per-language runtime enablement via setRuntimeEnablement', async () => {
